Revert "Bail early if dependency or kernel module fails to build (#5670)"
This reverts commit 0bf9f11953ba51cb979c4656c4f5d0700c7762df.

+++ mk/spksrc.spk.mk
@@ -599,11 +599,10 @@ kernel-modules-%:
 	   archs2process=$* ; \
 	fi ; \
 	$(MSG) ARCH to be processed: $${archs2process} ; \
-	set -e ; \
 	for arch in $${archs2process} ; do \
 	  $(MSG) "Processing $${arch} ARCH" ; \
 	  MAKEFLAGS= $(PSTAT_TIME) $(MAKE) WORK_DIR=$(PWD)/work-$* ARCH=$$(echo $${arch} | cut -f1 -d-) TCVERSION=$$(echo $${arch} | cut -f2 -d-) strip 2>&1 | tee --append build-$*.log ; \
-	  [ $${PIPESTATUS[0]} -eq 0 ] || false ; \
+	  [ $${PIPESTATUS[0]} -eq 0 ] || continue ; \
 	  $(MAKE) spkclean ; \
 	  rm -fr $(PWD)/work-$*/$(addprefix linux-, $${arch}) ; \
 	  $(MAKE) -C ../../toolchain/syno-$${arch} clean ; \
